--- modules/h2/README.md
@@ -1,3 +1,0 @@
-# H2 Extensions for Doobie
-
-This project provides (currently) a handful of H2-specific extensions.
