+++ modules/postgres/src/main/scala/doobie/contrib/postgresql/free/fastpath.scala
