+++ modules/postgres/src/main/scala/doobie/contrib/postgresql/hi/pgconnection.scala
