--- modules/postgresql/src/test/java/doobie/contrib/postgresql/MyJavaEnum.java
@@ -1,5 +1,0 @@
-package doobie.contrib.postgresql;
-
-public enum MyJavaEnum {
-  foo, bar;
-}
