+++ yax/postgres/src/test/java/doobie/postgres/MyJavaEnum.java
@@ -1,0 +1,5 @@
+package doobie.postgres;
+
+public enum MyJavaEnum {
+  foo, bar;
+}
